A0-539: Add more tests for the alerter (fka AZ-340) (#193)
The existing tests are integration-ish. This adds tests—some based on the
existing tests, some new—which test the alerter in a more functional way
(setting it up, giving it input, and checking its output).

These tests are more confined in what they test and generally simpler, while
the existing tests are more holistic and able to perform more checks (for
example, they can verify that behavior is correct with regard to RMCs). Both
sets of tests have value.

Additional changes:

* Remove extraneous data from AlerterResponse::AlertRequest

  `AlerterResponse` doesn't need a whole message in it, just the hash of the
  alert to request and the ID of the node to request it from. It's less confusing
  this way, too.

* Bump version to 0.13.1