Set opt-level to 3 in test (#244)
The crash recovery tests are flaky. It seems they sometimes overload the
machine with the 20 or so simulated workers producing messages faster
than they can process them. With opt-level=3 it doesn't seem to happen.
Compilation from a clean state is somewhat longer (46 seconds vs 18
seconds in the one test I did), but you don't often recompile all the
dependencies so this shouldn't be such an issue.

I also reenabled "medium" tests on CI - I hope they will pass now and
it's misleading to have them in the repo but not run on CI anyway.