A0-1045: Rewrite network in ordering example to UDP (#246)
* Rewrite network in ordering example to UDP

This solves the problem of establishing a connection with every network
send which seemed to blow up the system's capability to make more
connections in some circumstances. At the same time, the mock network
remains simple with no tracking of open connections and reconnection
attempts. A potential downside is that the UDP datagrams may be lost,
however this should rarely happen locally and the protocol should be
resistant to that in any case.

* Comment Box usage